Innfører LRU cache for html-endepunkter
