+++ .gitignore
@@ -15,7 +15,7 @@ npm-debug.log*
 .env.local
 .env.development.local
 .env.test.local
-.env.production.local
+.env
 /server/.env
 /server/.env.local
 /server/.env.development.local
