+++ server/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "dotenv": "16.0.3",
     "express": "4.18.2",
-    "node-cache": "5.1.2",
+    "lru-cache": "7.14.1",
     "node-fetch": "2.6.9",
     "node-schedule": "2.1.1",
     "lodash.debounce": "4.0.8",
@@ -20,6 +20,7 @@
   },
   "devDependencies": {
     "@types/express": "4.17.17",
+    "@types/lru-cache": "7.10.10",
     "@types/node": "18.11.18",
     "@types/node-schedule": "2.1.0",
     "concurrently": "7.6.0",
